+++ src/main/java/rocks/friedrich/engine_omega/Game.java
@@ -228,18 +228,18 @@ public final class Game
         // Finde Klick auf Zeichenebene, die Position relativ zum Ursprung des
         // RenderPanel-Canvas.
         // Mausklick-Position muss mit Zoom-Wert verrechnet werden
-        float zoom = scene.getCamera().getZoom();
-        float rotation = scene.getCamera().getRotation();
+        double zoom = scene.getCamera().getZoom();
+        double rotation = scene.getCamera().getRotation();
         Vector position = scene.getCamera().getPosition();
         return new Vector(
-                position.getX() + (((float) Math.cos(Math.toRadians(rotation))
+                position.getX() + (((double) Math.cos(Math.toRadians(rotation))
                         * (mousePosition.x - width / 2f)
-                        + (float) Math.sin(Math.toRadians(rotation))
+                        + (double) Math.sin(Math.toRadians(rotation))
                                 * (mousePosition.y - height / 2f)))
                         / zoom,
-                position.getY() + (((float) Math.sin(Math.toRadians(rotation))
+                position.getY() + (((double) Math.sin(Math.toRadians(rotation))
                         * (mousePosition.x - width / 2f)
-                        - (float) Math.cos(Math.toRadians(rotation))
+                        - (double) Math.cos(Math.toRadians(rotation))
                                 * (mousePosition.y - height / 2f)))
                         / zoom);
     }
@@ -254,7 +254,7 @@ public final class Game
             java.awt.event.MouseWheelEvent mouseWheelEvent)
     {
         MouseWheelEvent mouseWheelAction = new MouseWheelEvent(
-                (float) mouseWheelEvent.getPreciseWheelRotation());
+                (double) mouseWheelEvent.getPreciseWheelRotation());
         gameLogic.enqueue(
                 () -> scene.invokeMouseWheelMoveListeners(mouseWheelAction));
     }
